refactor(ui): dedupe admin Skeleton into shared components/ui/skeleton
