+++ src/apps/admin/components/song-detail-panel/SongDetailPanelActions.tsx
@@ -11,7 +11,7 @@ import {
   MusicNote,
   UserMinus,
 } from "@phosphor-icons/react";
-import { Skeleton } from "./Skeleton";
+import { Skeleton } from "@/components/ui/skeleton";
 import type { SongDetailPanelViewModel } from "./useSongDetailPanel";
 
 type Props = Pick<
