--- src/apps/admin/components/user-profile-panel/Skeleton.tsx
@@ -1,5 +1,0 @@
-import { cn } from "@/lib/utils";
-
-export const Skeleton = ({ className }: { className?: string }) => (
-  <div className={cn("bg-os-panel-bg animate-pulse rounded", className)} />
-);
