+++ src/apps/admin/components/user-profile-panel/UserProfilePanelRoomsSection.tsx
@@ -1,5 +1,5 @@
 import { SectionHeader } from "./SectionHeader";
-import { Skeleton } from "./Skeleton";
+import { Skeleton } from "@/components/ui/skeleton";
 import type { UserProfilePanelViewModel } from "./useUserProfilePanel";
 
 type Props = Pick<
